refactor(#50): User datasource 위치 변경
