+++ app/src/main/java/com/ssafy/tott/data/repository/UserRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.ssafy.tott.data.repository
 
-import com.ssafy.tott.data.datasource.UserDataSource
+import com.ssafy.tott.data.datasource.remote.UserRemoteDataSource
 import com.ssafy.tott.data.datasource.local.UserTokenDataSource
 import com.ssafy.tott.data.model.request.SignupRequest.Companion.toData
 import com.ssafy.tott.data.model.response.AuthTokenRemoteResponse
@@ -12,19 +12,19 @@ import kotlinx.coroutines.flow.map
 import javax.inject.Inject
 
 class UserRepositoryImpl @Inject constructor(
-    private val userDataSource: UserDataSource,
+    private val userRemoteDataSource: UserRemoteDataSource,
     private val userTokenDataSource: UserTokenDataSource
 ) : UserRepository {
     override fun makeCertNum(registerUser: RegisterUser): Flow<Result<Unit>> {
-        return userDataSource.makeCertNum(registerUser.toData())
+        return userRemoteDataSource.makeCertNum(registerUser.toData())
     }
 
     override fun requestCert(accountNum: String, certNum: String): Flow<Result<Unit>> {
-        return userDataSource.requestCert(accountNum, certNum)
+        return userRemoteDataSource.requestCert(accountNum, certNum)
     }
 
     override fun login(id: String, password: String): Flow<Result<AuthToken>> {
-        return userDataSource.login(id, password).map { result ->
+        return userRemoteDataSource.login(id, password).map { result ->
             result.map(AuthTokenRemoteResponse::toDomain)
         }
     }
